feat: adjust structure table

+++ database/migrations/2025_03_27_130232_create_car_brands_table.php
@@ -7,7 +7,7 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up() {
         Schema::create('car_brands', function (Blueprint $table) {
-            $table->unsignedSmallInteger('id')->autoIncrement()->primary();
+            $table->unsignedTinyInteger('id')->autoIncrement()->primary();
             $table->string('name', 100)->unique();
             $table->timestamps();
         });
