+++ database/migrations/2025_03_28_130232_create_cars_table.php
@@ -11,17 +11,17 @@ return new class extends Migration
         Schema::create('cars', function (Blueprint $table) {
             $table->unsignedSmallInteger('id')->autoIncrement()->primary();
 
-            $table->unsignedSmallInteger('brand_id');
-            $table->unsignedSmallInteger('fuel_type_id');
-            $table->unsignedSmallInteger('car_type_id');
-            $table->unsignedSmallInteger('transmission_type_id');
+            $table->unsignedTinyInteger('brand_id');
+            $table->unsignedTinyInteger('fuel_type_id');
+            $table->unsignedTinyInteger('car_type_id');
+            $table->unsignedTinyInteger('transmission_type_id');
 
             $table->string('name', 100);
             $table->text('image_name')->nullable();
-            $table->unsignedSmallInteger('price');
+            $table->unsignedInteger('price');
             $table->year('manufacture_year');
             $table->foreign('brand_id')->references('id')->on('car_brands')->onUpdate('cascade')->onDelete('cascade');
-            $table->unsignedSmallInteger('mileage');
+            $table->unsignedInteger('mileage');
             $table->foreign('fuel_type_id')->references('id')->on('fuel_types')->onUpdate('cascade')->onDelete('cascade');
             $table->unsignedSmallInteger('engine_capacity');
             $table->foreign('car_type_id')->references('id')->on('car_types')->onUpdate('cascade')->onDelete('cascade');
